requirements.txt
streamlit
python-dotenv
google-generativeai
Pillow
gTTS
google-genai
